[tcp_datagram_socket] Use uint16_t in send function for message size
- This better reflects how big a message can really be

+++ include/liblinuxpp/net/tcp_datagram_socket.hpp
@@ -192,7 +192,7 @@ namespace net
          *  @param flags The flags to pass to sendto see man 2 sendto
          */
         std::size_t send(void const * buf,
-                         const std::size_t length,
+                         const std::uint16_t length,
                          const int flags = 0);
 
         /** Calls sendmsg
